Fix Android build and update build scripts
Fixes:
- FfiError field renamed from `message` to `reason` to avoid
  conflict with Kotlin Throwable.message
- block.depth converted to Int for Compose dp calculation
- compileSdk reduced from 36 to 35 (AGP 8.7.3 compatibility)

Build improvements:
- build-android.sh now generates gradle wrapper if missing and
  builds the APK
- run-apk-adb-emulator.sh updated for new Kotlin project
- Gradle wrapper files added to .gitignore (generated, not committed)

Prompts:
- "I don't want to use android studio. what do i need to get an adb onto the emulator? update run-apk-adb-emulator.sh till it works"
- "shouldn't we script this rather than running it one off? everything needs to be repeatable"
- "run it again"
- "commit? a passing build seems like progress"

+++ android/app/src/main/java/co/rustworkshop/markdownneuraxis/MainActivity.kt
@@ -245,7 +245,7 @@ fun FileViewScreen(
 
 @Composable
 fun RenderBlock(block: RenderBlockDto) {
-    val indent = (block.depth * 16).dp
+    val indent = (block.depth.toInt() * 16).dp
 
     when (block.kind) {
         "heading" -> {
